requirements.txt
accelerate==0.25.0
adal==1.2.7
aiofiles==23.1.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.6.0
anthropic==0.10.0
antlr4-python3-runtime==4.9.3
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asitop==0.0.24
astor==0.8.1
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
azure-cognitiveservices-search-websearch==2.0.0
azure-common==1.1.28
azure-core==1.29.1
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.41.0
bleach==4.1.0
blessed==1.20.0
blinker==1.6.2
boto3==1.34.25
botocore==1.34.25
brotlipy==0.7.0
bs4==0.0.2
build==1.0.3
cachetools==5.3.1
certifi==2023.11.17
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.6
coloredlogs==15.0.1
comm==0.1.2
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
dashing==0.1.0
dataclasses-json==0.6.3
datasets==2.15.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dirtyjson==1.0.8
distro==1.9.0
duckdb==0.9.2
effdet==0.4.1
einops==0.7.0
emoji==2.9.0
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==0.8.3
Faker==19.13.0
fastapi==0.101.0
fastjsonschema==2.16.2
ffmpy==0.3.1
filelock==3.12.2
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.34
google-auth==2.28.0
googleapis-common-protos==1.62.0
gradio==3.39.0
gradio_client==0.3.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
hf_transfer==0.1.4
html2text==2020.1.16
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
InstructorEmbedding==1.0.1
iopath==0.1.10
ipykernel==6.19.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
json5==0.9.6
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langdetect==1.0.9
langsmith==0.0.77
layoutparser==0.3.4
linkify-it-py==2.0.2
llama-hub==0.0.74
llama-index==0.10.10
llama-index-agent-openai==0.1.4
llama-index-cli==0.1.3
llama-index-core==0.10.10
llama-index-embeddings-huggingface==0.1.1
llama-index-embeddings-instructor==0.1.1
llama-index-embeddings-openai==0.1.5
llama-index-indices-managed-llama-cloud==0.1.2
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.5
llama-index-multi-modal-llms-openai==0.1.3
llama-index-packs-subdoc-summary==0.1.1
llama-index-program-openai==0.1.3
llama-index-question-gen-openai==0.1.2
llama-index-readers-file==0.1.4
llama-index-readers-llama-parse==0.1.2
llama-index-vector-stores-chroma==0.1.2
llama-parse==0.3.4
llamaindex-py-client==0.1.13
llmlingua==0.1.5
loguru==0.7.2
lxml==5.0.0
Markdown==3.5.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
milvus==2.2.12
minio==7.2.4
mistune==0.8.4
mlx==0.5.0
mlx-graphs==0.0.4
mlx-lm==0.0.14
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.1
openai==1.6.1
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.2
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandasai==1.5.17
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
peft==0.5.0.dev0
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.11.2
Pillow==10.1.0
pip==23.2.1
platformdirs==2.5.2
ply==3.11
portalocker==2.8.2
posthog==3.4.2
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyaml==23.12.0
pyarrow==12.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.4
pydantic_core==2.14.6
pydeck==0.8.0
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
pymilvus==2.3.6
Pympler==1.0.1
PyMuPDF==1.23.24
PyMuPDFb==1.23.22
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pypandoc==1.12
pyparsing==3.0.9
pypdf==4.0.2
PyPDF2==3.0.1
pypdfium2==4.25.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-iso639==2024.1.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.23
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.4.2
QtPy==2.2.0
rapidfuzz==3.6.1
redis==5.0.1
referencing==0.32.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.16.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.4
SQLAlchemy==1.4.50
stack-data==0.2.0
starlette==0.27.0
streamlit==1.31.1
streamlit-feedback==0.1.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.2
tqdm==4.66.1
traitlets==5.7.1
transformers==4.38.1
transformers-stream-generator==0.0.4
trl==0.4.7
trubrics==1.6.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
ujson==5.8.0
unstructured==0.11.8
unstructured-client==0.15.1
unstructured-inference==0.7.24
unstructured.pytesseract==0.3.12
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==0.58.0
websockets==11.0.3
wheel==0.38.4
widgetsnbextension==4.0.5
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
